Rename FavoriteModel to ArtPieceModel
