+++ src/main/java/com/arte/backend/model/artpiece/ArtPieceModel.java
@@ -1,4 +1,4 @@
-package com.arte.backend.model.favorites;
+package com.arte.backend.model.artpiece;
 
 import com.arte.backend.model.apiresponse.WebImage;
 import lombok.AllArgsConstructor;
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 @Builder
-public class FavoritesModel {
+public class ArtPieceModel {
     private String longTitle;
     private WebImage webImage;
     private String objectNumber;
