+++ src/main/java/com/arte/backend/util/helper/Converter.java
@@ -1,17 +1,17 @@
 package com.arte.backend.util.helper;
 
 import com.arte.backend.model.favorites.FavoriteFolderModel;
-import com.arte.backend.model.favorites.FavoritesModel;
+import com.arte.backend.model.artpiece.ArtPieceModel;
 
 import java.util.*;
 import java.util.stream.Collectors;
 
 public class Converter {
-    public static List<FavoritesModel> favoritesModelSetToSortedListByTitle(Set<FavoritesModel> inputSet) {
+    public static List<ArtPieceModel> favoritesModelSetToSortedListByTitle(Set<ArtPieceModel> inputSet) {
         if (inputSet == null) { return null; }
-        List<FavoritesModel> resultList = new ArrayList<>(inputSet);
+        List<ArtPieceModel> resultList = new ArrayList<>(inputSet);
         resultList = resultList.stream()
-                .sorted(Comparator.comparing(FavoritesModel::getLongTitle))
+                .sorted(Comparator.comparing(ArtPieceModel::getLongTitle))
                 .collect(Collectors.toList());
         return resultList;
     }
